updated from main

+++ Assets/Scripts/Difficulty.cs
@@ -1,4 +1,4 @@
 public class Difficulty : Singleton<Difficulty>
 {
-    public int difficulty;
+    public float difficulty;
 }
